mail-parser/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
openai==1.3.7
python-multipart==0.0.6
email-validator==2.1.0
imapclient==2.3.1